lang/programming/python/studioclassroom_FulltextSearch/template/requirements.txt
Flask
Flask-Cors
chardet
mecab-python3
unidic-lite
xmltodict
psycopg2-binary
pymysql==0.9.3
xmltodict
zhconv